Fix bug in attachment mapping for existing attachments
Also fixed a bug that exists in the current wordpress importer plugin where the attachment_url does not match the url in the content. We should also try and use the GUID and strip the extension as we do for the attachment_url. This is not an issue when only guid is defined but that is not always the case.